Add files via upload
In my project, I'm going to train a K-Nearest Neighbor (KNN) model with different values of K (2, 4, 5, 6, 7, 10) and evaluate how the model performs with various train-test splits (60:40, 70:30, 75:25, 80:20, 90:10, 95:5). I'll be assessing the accuracy and confusion matrix for 36 different scenarios to analyze how the model's performance depends on both the value of K and the train-test split ratio.